[refactor] book info entity 위치 변경에 따른 수정
