+++ BE/src/main/java/com/FTIsland/BE/repository/BookInfoRepository.java
@@ -1,7 +1,6 @@
 package com.FTIsland.BE.repository;
 
-import com.FTIsland.BE.entity.BookInfoEntity;
-import com.google.api.services.storage.Storage;
+import com.FTIsland.BE.island.entity.BookInfoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
